+++ src/_main.ts
@@ -1,9 +1,18 @@
 
-import { loadFromArrayBuffer, loadFromDataUri, loadFromUrl } from './loader';
+import {
+	config, loadAsyl, loadFromArrayBuffer, loadFromDataUri, loadFromUrl,
+	memory, dataView, asyl
+} from './loader';
 
 globalThis.asyl =
 {
+	config,
+	loadAsyl,
 	loadFromArrayBuffer,
 	loadFromDataUri,
-	loadFromUrl
+	loadFromUrl,
+
+	memory,
+	dataView,
+	asyl
 };