+++ src/main.ts
@@ -1,4 +1,8 @@
 
-export { loadFromArrayBuffer, loadFromDataUri, loadFromUrl } from './loader';
+export {
+	config, loadFromArrayBuffer, loadFromDataUri, loadFromUrl, loadAsyl,
+	memory, dataView, asyl
+} from './loader';
+
 export { default as Module } from './module';
 export { default as wasi } from './wasi';